refactor: split telemetry snapshot projection
