+++ docs/CODEBASE_AUDIT_ACTION_TABLE_2026-03-09.md
@@ -39,6 +39,7 @@ This table captures every issue found in this pass. It is broad and high-signal,
 | R19 | Fixed | Daemon loop decomposition | `internal/daemon/server.go`, `internal/daemon/server_loops.go` | Collection, spool, hook-spool, retention, and poll loops no longer live inline in the main daemon server file. | Continue splitting by loop family if the new file grows too large. |
 | R20 | Fixed | TUI model/client/provider parsing duplication | `internal/core/usage_breakdowns.go`, `internal/tui/tiles_composition.go` | Model, client, provider, upstream-provider, and interface-client aggregation/parsing now live in shared core extractors, leaving the TUI composition layer as a thin adapter over typed breakdown entries. | The remaining TUI parsing drift is now mostly in analytics/detail-specific sections rather than the main composition bars. |
 | R21 | Fixed | OpenRouter provider-resolution split | `internal/providers/openrouter/openrouter.go`, `internal/providers/openrouter/provider_resolution.go` | Hosting-provider resolution, BYOK cost inference, and provider-name heuristics moved out of the main OpenRouter provider file into a dedicated helper unit. | Continue splitting analytics/generation pagination/projection concerns. |
+| R22 | Fixed | Telemetry snapshot projection split | `internal/telemetry/usage_view.go`, `internal/telemetry/usage_view_projection.go` | Snapshot projection, stale-metric cleanup, daily-series projection, and windowed metric emission moved out of the main usage-view file into a dedicated projection unit. | Continue with the same split for the SQL/query layer. |
 
 ## Action Table
 
@@ -49,7 +50,7 @@ This table captures every issue found in this pass. It is broad and high-signal,
 | A3 | P2 | UI metric-prefix parsing | `internal/tui/analytics.go:663-729`, `internal/tui/detail.go`, `internal/core/usage_breakdowns.go` | The main composition bars now consume shared extractors, but analytics/detail-specific sections still decode some raw metric-key conventions directly in TUI code. | Continue promoting the remaining analytics/detail extractors into `internal/core` or `internal/telemetry` so renderers consume structured sections instead of re-parsing maps. | Removes a large class of UI drift bugs and reduces per-render work. |
 | A4 | P1 | OpenRouter provider size | `internal/providers/openrouter/openrouter.go`, `internal/providers/openrouter/provider_resolution.go` | `openrouter.go` is smaller after the provider-resolution split, but it still mixes auth probing, credits, keys, analytics parsing, generation pagination, metadata enrichment, and output projection in one large file. | Continue splitting into `api_client`, `analytics`, `generations`, `projection`, and `types` units. | Easier maintenance, smaller diff surface, faster targeted testing. |
 | A5 | P1 | Cursor provider responsibility overload | `internal/providers/cursor/cursor.go:181-335`, `internal/providers/cursor/cursor.go:903-1006`, `internal/providers/cursor/cursor.go:1087-2086` | Cursor provider combines API orchestration, local SQLite readers, token extraction, and two independent caches in one class. | Split into `api`, `trackingdb`, `statedb`, `cache`, and `snapshot_projection` modules. Move token extraction out of provider hot path. | Cleaner boundaries and less risk of local/API logic regressions. |
-| A6 | P1 | Telemetry usage-view monolith | `internal/telemetry/usage_view.go`, `internal/telemetry/usage_view_helpers.go` | `usage_view.go` is smaller now, but it is still simultaneously query planner, SQL execution layer, aggregation engine, naming normalizer, and snapshot projection layer. | Split into `query_*`, `aggregate_*`, `projection_*`, and `mcp_*` units. Add a typed intermediate aggregation model. | Easier optimization and safer incremental changes. |
+| A6 | P1 | Telemetry usage-view monolith | `internal/telemetry/usage_view.go`, `internal/telemetry/usage_view_helpers.go`, `internal/telemetry/usage_view_projection.go` | `usage_view.go` is materially smaller after the helper/projection split, but it still mixes query planning, SQL execution, aggregation orchestration, and query-specific shaping logic. | Continue splitting into `query_*`, `aggregate_*`, and remaining orchestration units. Add a typed intermediate aggregation model. | Easier optimization and safer incremental changes. |
 | A7 | P2 | Daemon service monolith | `internal/daemon/server.go`, `internal/daemon/server_http.go`, `internal/daemon/server_read_model.go`, `internal/daemon/server_loops.go` | The daemon is materially less coupled after the logging/cache/http/loop split, but polling, collection, retention, and spool maintenance still share one large runtime helper unit. | Continue splitting the loop-heavy runtime into `polling`, `collection`, `retention`, and `spool` units. | Lower mental load and easier concurrency review. |
 | A11 | P2 | Time-dependent logic without injectable clock | `internal/providers/openrouter/openrouter.go:728`, `internal/providers/ollama/ollama.go:1088`, `internal/core/analytics_normalize.go:61-103` | Cursor’s main time-sensitive path now uses an injectable clock, but several other providers and analytics helpers still read `time.Now()` directly, often mixing local time and UTC. | Extend the clock abstraction to the remaining provider and analytics subsystems and standardize UTC/local semantics per provider. | Better determinism and fewer timezone edge cases. |
 | A12 | P2 | Test file sprawl and fixture duplication | `internal/providers/openrouter/openrouter_test.go`, `internal/providers/copilot/copilot_test.go`, `internal/telemetry/usage_view_test.go`, `internal/config/config_test.go` | Some tests are 1000-2600 LOC and re-encode similar setup logic inline. They are valuable but expensive to navigate and update. | Extract fixture builders and scenario helpers per package. Keep top-level tests declarative. | Faster iteration and simpler maintenance of large test suites. |
