+++ internal/telemetry/usage_view.go
@@ -237,302 +237,6 @@ func applyCanonicalUsageViewWithDB(
 	return out, nil
 }
 
-func applyUsageViewToSnapshot(snap *core.UsageSnapshot, agg *telemetryUsageAgg, timeWindow core.TimeWindow) {
-	if snap == nil || agg == nil {
-		return
-	}
-	authoritativeCost := usageAuthoritativeCost(*snap)
-	windowLabel := string(timeWindow)
-	snap.EnsureMaps()
-	if snap.DailySeries == nil {
-		snap.DailySeries = make(map[string][]core.TimePoint)
-	}
-
-	// Save API-sourced model cost metrics (billing-cycle) before cleanup.
-	// These will be restored if telemetry events lack sufficient cost attribution.
-	savedAPIModelCosts := make(map[string]core.Metric)
-	for key, m := range snap.Metrics {
-		if strings.HasPrefix(key, "model_") && strings.HasSuffix(key, "_cost") && m.Window == "billing-cycle" {
-			savedAPIModelCosts[key] = m
-		}
-	}
-
-	metricsBefore := len(snap.Metrics)
-	_, hadFiveHourBefore := snap.Metrics["usage_five_hour"]
-	stripAllTime := timeWindow != "" && timeWindow != "all"
-	deletedCount := 0
-	for key, m := range snap.Metrics {
-		if strings.HasPrefix(key, "source_") ||
-			strings.HasPrefix(key, "client_") ||
-			strings.HasPrefix(key, "tool_") ||
-			strings.HasPrefix(key, "model_") ||
-			strings.HasPrefix(key, "project_") ||
-			strings.HasPrefix(key, "provider_") ||
-			strings.HasPrefix(key, "lang_") ||
-			strings.HasPrefix(key, "interface_") ||
-			isStaleActivityMetric(key) {
-			delete(snap.Metrics, key)
-			deletedCount++
-		} else if stripAllTime && m.Window == "all-time" && !isCurrentStateMetric(key) {
-			// Strip cumulative "all-time" metrics from the limit_snapshot so the
-			// TUI doesn't show misleading all-time counts under a windowed badge.
-			// Current-state metrics (plan quotas, billing, team) are preserved.
-			delete(snap.Metrics, key)
-			deletedCount++
-		}
-	}
-	_, hasFiveHourAfter := snap.Metrics["usage_five_hour"]
-	core.Tracef("[usage_view] %s: cleanup deleted %d/%d metrics, usage_five_hour before=%v after=%v",
-		snap.ProviderID, deletedCount, metricsBefore, hadFiveHourBefore, hasFiveHourAfter)
-	telemetryPrefixes := []string{"source_", "client_", "tool_", "model_", "project_", "provider_", "usage_", "analytics_"}
-	extendedPrefixes := append(telemetryPrefixes, "lang_", "jsonl_")
-	deleteByPrefixes(snap.Raw, extendedPrefixes)
-	deleteByPrefixes(snap.Attributes, telemetryPrefixes)
-	deleteByPrefixes(snap.Diagnostics, telemetryPrefixes)
-	for key := range snap.DailySeries {
-		if strings.HasPrefix(key, "usage_model_") ||
-			strings.HasPrefix(key, "usage_source_") ||
-			strings.HasPrefix(key, "usage_project_") ||
-			strings.HasPrefix(key, "usage_client_") ||
-			strings.HasPrefix(key, "tokens_client_") ||
-			key == "analytics_cost" ||
-			key == "analytics_requests" ||
-			key == "analytics_tokens" {
-			delete(snap.DailySeries, key)
-		}
-	}
-
-	// Replace stale template ModelUsage with time-windowed records from
-	// telemetry. The template's ModelUsage represents the full billing cycle
-	// and would be misleading for shorter time windows.
-	snap.ModelUsage = nil
-	modelCostTotal := 0.0
-	for _, model := range agg.Models {
-		mk := sanitizeMetricID(model.Model)
-		snap.Metrics["model_"+mk+"_input_tokens"] = core.Metric{Used: core.Float64Ptr(model.InputTokens), Unit: "tokens", Window: windowLabel}
-		snap.Metrics["model_"+mk+"_output_tokens"] = core.Metric{Used: core.Float64Ptr(model.OutputTokens), Unit: "tokens", Window: windowLabel}
-		snap.Metrics["model_"+mk+"_cached_tokens"] = core.Metric{Used: core.Float64Ptr(model.CachedTokens), Unit: "tokens", Window: windowLabel}
-		snap.Metrics["model_"+mk+"_reasoning_tokens"] = core.Metric{Used: core.Float64Ptr(model.Reasoning), Unit: "tokens", Window: windowLabel}
-		snap.Metrics["model_"+mk+"_cost_usd"] = core.Metric{Used: core.Float64Ptr(model.CostUSD), Unit: "USD", Window: windowLabel}
-		snap.Metrics["model_"+mk+"_requests"] = core.Metric{Used: core.Float64Ptr(model.Requests), Unit: "requests", Window: windowLabel}
-		snap.Metrics["model_"+mk+"_requests_today"] = core.Metric{Used: core.Float64Ptr(model.Requests1d), Unit: "requests", Window: "1d"}
-		modelCostTotal += model.CostUSD
-		snap.ModelUsage = append(snap.ModelUsage, core.ModelUsageRecord{
-			RawModelID:      model.Model,
-			RawSource:       "telemetry",
-			Window:          windowLabel,
-			InputTokens:     core.Float64Ptr(model.InputTokens),
-			OutputTokens:    core.Float64Ptr(model.OutputTokens),
-			CachedTokens:    core.Float64Ptr(model.CachedTokens),
-			ReasoningTokens: core.Float64Ptr(model.Reasoning),
-			TotalTokens:     core.Float64Ptr(model.TotalTokens),
-			CostUSD:         core.Float64Ptr(model.CostUSD),
-			Requests:        core.Float64Ptr(model.Requests),
-		})
-	}
-	// When telemetry events lack cost attribution but the provider's API
-	// supplied per-model cost data (e.g. Cursor's GetAggregatedUsageEvents),
-	// restore the API model costs so the Model Burn section shows the real
-	// per-model breakdown instead of a single "unattributed" entry.
-	telemetryCostInsufficient := authoritativeCost > 0 && modelCostTotal < authoritativeCost*0.1
-	if telemetryCostInsufficient && len(savedAPIModelCosts) > 0 {
-		for key, m := range savedAPIModelCosts {
-			snap.Metrics[key] = m
-		}
-		core.Tracef("[usage_view] %s: restored %d API model cost metrics (telemetry cost %.2f << authoritative %.2f)",
-			snap.ProviderID, len(savedAPIModelCosts), modelCostTotal, authoritativeCost)
-	} else if len(agg.Models) > 0 {
-		// Only compute unattributed model cost when telemetry has meaningful
-		// cost data. When agg.Models is empty (no events in window), the
-		// authoritativeCost represents the full billing cycle — attributing
-		// it as "unattributed" would be misleading for the selected time range.
-		if delta := authoritativeCost - modelCostTotal; authoritativeCost > 0 && delta > 0.000001 {
-			uk := "model_unattributed"
-			snap.Metrics[uk+"_cost_usd"] = core.Metric{Used: core.Float64Ptr(delta), Unit: "USD", Window: windowLabel}
-			snap.SetDiagnostic("telemetry_unattributed_model_cost_usd", fmt.Sprintf("%.6f", delta))
-		}
-	}
-
-	if !strings.EqualFold(strings.TrimSpace(snap.ProviderID), "codex") {
-		providerCostTotal := 0.0
-		for _, provider := range agg.Providers {
-			pk := sanitizeMetricID(provider.Provider)
-			snap.Metrics["provider_"+pk+"_cost_usd"] = core.Metric{Used: core.Float64Ptr(provider.CostUSD), Unit: "USD", Window: windowLabel}
-			snap.Metrics["provider_"+pk+"_input_tokens"] = core.Metric{Used: core.Float64Ptr(provider.Input), Unit: "tokens", Window: windowLabel}
-			snap.Metrics["provider_"+pk+"_output_tokens"] = core.Metric{Used: core.Float64Ptr(provider.Output), Unit: "tokens", Window: windowLabel}
-			snap.Metrics["provider_"+pk+"_requests"] = core.Metric{Used: core.Float64Ptr(provider.Requests), Unit: "requests", Window: windowLabel}
-			providerCostTotal += provider.CostUSD
-		}
-		if delta := authoritativeCost - providerCostTotal; authoritativeCost > 0 && delta > 0.000001 {
-			uk := "provider_unattributed"
-			snap.Metrics[uk+"_cost_usd"] = core.Metric{Used: core.Float64Ptr(delta), Unit: "USD", Window: windowLabel}
-			snap.SetDiagnostic("telemetry_unattributed_provider_cost_usd", fmt.Sprintf("%.6f", delta))
-		}
-	}
-
-	for _, source := range agg.Sources {
-		sk := sanitizeMetricID(source.Source)
-		// Only emit source_*_requests_today (used by TUI's today-fallback path).
-		// source_*_requests is intentionally omitted: client_*_requests covers the
-		// same data, and emitting both causes the TUI to double-count requests due
-		// to Go's random map iteration order.
-		snap.Metrics["source_"+sk+"_requests_today"] = core.Metric{Used: core.Float64Ptr(source.Requests1d), Unit: "requests", Window: "1d"}
-
-		snap.Metrics["client_"+sk+"_total_tokens"] = core.Metric{Used: core.Float64Ptr(source.Tokens), Unit: "tokens", Window: windowLabel}
-		snap.Metrics["client_"+sk+"_input_tokens"] = core.Metric{Used: core.Float64Ptr(source.Input), Unit: "tokens", Window: windowLabel}
-		snap.Metrics["client_"+sk+"_output_tokens"] = core.Metric{Used: core.Float64Ptr(source.Output), Unit: "tokens", Window: windowLabel}
-		snap.Metrics["client_"+sk+"_cached_tokens"] = core.Metric{Used: core.Float64Ptr(source.Cached), Unit: "tokens", Window: windowLabel}
-		snap.Metrics["client_"+sk+"_reasoning_tokens"] = core.Metric{Used: core.Float64Ptr(source.Reasoning), Unit: "tokens", Window: windowLabel}
-		snap.Metrics["client_"+sk+"_requests"] = core.Metric{Used: core.Float64Ptr(source.Requests), Unit: "requests", Window: windowLabel}
-		snap.Metrics["client_"+sk+"_sessions"] = core.Metric{Used: core.Float64Ptr(source.Sessions), Unit: "sessions", Window: windowLabel}
-	}
-	for _, project := range agg.Projects {
-		pk := sanitizeMetricID(project.Project)
-		if pk == "" {
-			continue
-		}
-		snap.Metrics["project_"+pk+"_requests"] = core.Metric{Used: core.Float64Ptr(project.Requests), Unit: "requests", Window: windowLabel}
-		snap.Metrics["project_"+pk+"_requests_today"] = core.Metric{Used: core.Float64Ptr(project.Requests1d), Unit: "requests", Window: "1d"}
-	}
-
-	var totalToolCalls float64
-	var totalToolCallsOK float64
-	var totalToolCallsError float64
-	var totalToolCallsAborted float64
-	for _, tool := range agg.Tools {
-		tk := sanitizeMetricID(tool.Tool)
-		snap.Metrics["tool_"+tk] = core.Metric{Used: core.Float64Ptr(tool.Calls), Unit: "calls", Window: windowLabel}
-		snap.Metrics["tool_"+tk+"_today"] = core.Metric{Used: core.Float64Ptr(tool.Calls1d), Unit: "calls", Window: "1d"}
-		totalToolCalls += tool.Calls
-		totalToolCallsOK += tool.CallsOK
-		totalToolCallsError += tool.CallsError
-		totalToolCallsAborted += tool.CallsAborted
-	}
-	if totalToolCalls > 0 {
-		snap.Metrics["tool_calls_total"] = core.Metric{Used: core.Float64Ptr(totalToolCalls), Unit: "calls", Window: windowLabel}
-		snap.Metrics["tool_completed"] = core.Metric{Used: core.Float64Ptr(totalToolCallsOK), Unit: "calls", Window: windowLabel}
-		snap.Metrics["tool_errored"] = core.Metric{Used: core.Float64Ptr(totalToolCallsError), Unit: "calls", Window: windowLabel}
-		snap.Metrics["tool_cancelled"] = core.Metric{Used: core.Float64Ptr(totalToolCallsAborted), Unit: "calls", Window: windowLabel}
-		successRate := 0.0
-		if totalToolCalls > 0 {
-			successRate = (totalToolCallsOK / totalToolCalls) * 100
-		}
-		snap.Metrics["tool_success_rate"] = core.Metric{Used: core.Float64Ptr(successRate), Unit: "%", Window: windowLabel}
-	}
-
-	// MCP server metrics.
-	var mcpTotalCalls, mcpTotalCalls1d float64
-	for _, srv := range agg.MCPServers {
-		sk := sanitizeMetricID(srv.Server)
-		snap.Metrics["mcp_"+sk+"_total"] = core.Metric{Used: core.Float64Ptr(srv.Calls), Unit: "calls", Window: windowLabel}
-		snap.Metrics["mcp_"+sk+"_total_today"] = core.Metric{Used: core.Float64Ptr(srv.Calls1d), Unit: "calls", Window: "1d"}
-		mcpTotalCalls += srv.Calls
-		mcpTotalCalls1d += srv.Calls1d
-		for _, fn := range srv.Functions {
-			fk := sanitizeMetricID(fn.Function)
-			snap.Metrics["mcp_"+sk+"_"+fk] = core.Metric{Used: core.Float64Ptr(fn.Calls), Unit: "calls", Window: windowLabel}
-		}
-	}
-	if mcpTotalCalls > 0 {
-		snap.Metrics["mcp_calls_total"] = core.Metric{Used: core.Float64Ptr(mcpTotalCalls), Unit: "calls", Window: windowLabel}
-		snap.Metrics["mcp_calls_total_today"] = core.Metric{Used: core.Float64Ptr(mcpTotalCalls1d), Unit: "calls", Window: "1d"}
-		snap.Metrics["mcp_servers_active"] = core.Metric{Used: core.Float64Ptr(float64(len(agg.MCPServers))), Unit: "servers", Window: windowLabel}
-	}
-
-	for _, lang := range agg.Languages {
-		lk := sanitizeMetricID(lang.Language)
-		snap.Metrics["lang_"+lk] = core.Metric{Used: core.Float64Ptr(lang.Requests), Unit: "requests", Window: windowLabel}
-	}
-
-	// Emit windowed activity metrics.
-	act := agg.Activity
-	if act.Messages > 0 {
-		snap.Metrics["messages_today"] = core.Metric{Used: core.Float64Ptr(act.Messages), Unit: "messages", Window: windowLabel}
-	}
-	if act.Sessions > 0 {
-		snap.Metrics["sessions_today"] = core.Metric{Used: core.Float64Ptr(act.Sessions), Unit: "sessions", Window: windowLabel}
-	}
-	if act.ToolCalls > 0 {
-		snap.Metrics["tool_calls_today"] = core.Metric{Used: core.Float64Ptr(act.ToolCalls), Unit: "calls", Window: windowLabel}
-		snap.Metrics["7d_tool_calls"] = core.Metric{Used: core.Float64Ptr(act.ToolCalls), Unit: "calls", Window: windowLabel}
-	}
-	if act.InputTokens > 0 {
-		snap.Metrics["today_input_tokens"] = core.Metric{Used: core.Float64Ptr(act.InputTokens), Unit: "tokens", Window: windowLabel}
-	}
-	if act.OutputTokens > 0 {
-		snap.Metrics["today_output_tokens"] = core.Metric{Used: core.Float64Ptr(act.OutputTokens), Unit: "tokens", Window: windowLabel}
-	}
-	if act.TotalCost > 0 {
-		snap.Metrics["today_api_cost"] = core.Metric{Used: core.Float64Ptr(act.TotalCost), Unit: "USD", Window: windowLabel}
-	}
-
-	// Emit windowed code stats.
-	cs := agg.CodeStats
-	if cs.FilesChanged > 0 {
-		snap.Metrics["composer_files_changed"] = core.Metric{Used: core.Float64Ptr(cs.FilesChanged), Unit: "files", Window: windowLabel}
-	}
-	if cs.LinesAdded > 0 {
-		snap.Metrics["composer_lines_added"] = core.Metric{Used: core.Float64Ptr(cs.LinesAdded), Unit: "lines", Window: windowLabel}
-	}
-	if cs.LinesRemoved > 0 {
-		snap.Metrics["composer_lines_removed"] = core.Metric{Used: core.Float64Ptr(cs.LinesRemoved), Unit: "lines", Window: windowLabel}
-	}
-
-	// Emit window-level aggregate metrics for the TUI header/tile display.
-	var windowRequests, windowCost, windowTokens float64
-	for _, model := range agg.Models {
-		windowRequests += model.Requests
-		windowCost += model.CostUSD
-		windowTokens += model.TotalTokens
-	}
-	if windowRequests > 0 {
-		snap.Metrics["window_requests"] = core.Metric{Used: core.Float64Ptr(windowRequests), Unit: "requests", Window: windowLabel}
-	}
-	if windowCost > 0 {
-		snap.Metrics["window_cost"] = core.Metric{Used: core.Float64Ptr(windowCost), Unit: "USD", Window: windowLabel}
-	}
-	if windowTokens > 0 {
-		snap.Metrics["window_tokens"] = core.Metric{Used: core.Float64Ptr(windowTokens), Unit: "tokens", Window: windowLabel}
-	}
-
-	snap.DailySeries["analytics_cost"] = pointsFromDaily(agg.Daily, func(v telemetryDayPoint) float64 { return v.CostUSD })
-	snap.DailySeries["analytics_requests"] = pointsFromDaily(agg.Daily, func(v telemetryDayPoint) float64 { return v.Requests })
-	snap.DailySeries["analytics_tokens"] = pointsFromDaily(agg.Daily, func(v telemetryDayPoint) float64 { return v.Tokens })
-	// Fixed-window cost metrics (7d_api_cost, 5h_block_cost, all_time_api_cost,
-	// usage_daily, usage_weekly) are preserved from the provider template — they
-	// come from the provider's Fetch() with real API data. We do NOT re-emit
-	// them here because agg.Daily is already filtered to the selected time
-	// window, so usageCostWindowsUTC would produce incorrect values (e.g.
-	// "7d cost" would equal "3d cost" when the user picks a 3-day window).
-
-	for model, series := range agg.ModelDaily {
-		snap.DailySeries["usage_model_"+sanitizeMetricID(model)] = series
-	}
-	for source, series := range agg.SourceDaily {
-		snap.DailySeries["usage_source_"+sanitizeMetricID(source)] = series
-	}
-	for project, series := range agg.ProjectDaily {
-		snap.DailySeries["usage_project_"+sanitizeMetricID(project)] = series
-	}
-	for client, series := range agg.ClientDaily {
-		snap.DailySeries["usage_client_"+sanitizeMetricID(client)] = series
-	}
-	for client, series := range agg.ClientTokens {
-		snap.DailySeries["tokens_client_"+sanitizeMetricID(client)] = series
-	}
-
-	snap.SetAttribute("telemetry_view", "canonical")
-	snap.SetAttribute("telemetry_source_of_truth", "canonical_usage_events")
-	snap.SetAttribute("telemetry_last_event_at", agg.LastOccurred)
-	if strings.TrimSpace(agg.Scope) != "" {
-		snap.SetAttribute("telemetry_scope", agg.Scope)
-	}
-	if strings.TrimSpace(agg.AccountID) != "" {
-		snap.SetAttribute("telemetry_scope_account_id", agg.AccountID)
-	}
-	snap.SetDiagnostic("telemetry_event_count", fmt.Sprintf("%d", agg.EventCount))
-}
-
 // queryTelemetryActiveProviders returns the set of provider IDs that have at least
 // one telemetry event in the database, regardless of time window. This is used to
 // distinguish providers that have a telemetry adapter (but may have no events in the
@@ -1534,105 +1238,6 @@ func normalizeProviderIDs(in []string) []string {
 	result := lo.Uniq(lo.Compact(normalized))
 	sort.Strings(result)
 	return result
-}
-
-func pointsFromDaily(in []telemetryDayPoint, pick func(telemetryDayPoint) float64) []core.TimePoint {
-	return lo.Map(in, func(row telemetryDayPoint, _ int) core.TimePoint {
-		return core.TimePoint{Date: row.Day, Value: pick(row)}
-	})
-}
-
-// isStaleActivityMetric returns true for metrics that are computed by the provider
-// with hardcoded time windows (today/7d/all-time) and should be replaced by
-// telemetry-windowed equivalents.
-func isStaleActivityMetric(key string) bool {
-	// Activity counters with hardcoded time windows.
-	switch key {
-	case "messages_today", "sessions_today", "tool_calls_today",
-		"7d_tool_calls", "all_time_tool_calls", "tool_calls_total",
-		"tool_completed", "tool_errored", "tool_cancelled", "tool_success_rate",
-		"today_input_tokens", "today_output_tokens",
-		"7d_input_tokens", "7d_output_tokens",
-		"all_time_input_tokens", "all_time_output_tokens",
-		"all_time_cache_read_tokens", "all_time_cache_create_tokens",
-		"all_time_cache_create_5m_tokens", "all_time_cache_create_1h_tokens",
-		"all_time_reasoning_tokens",
-		"today_api_cost",
-		"burn_rate",
-		"composer_lines_added", "composer_lines_removed",
-		"composer_files_changed":
-		return true
-	}
-	// Fixed-window cost metrics from provider Fetch() are preserved —
-	// the telemetry view does NOT re-emit them (it only has windowed data).
-	switch key {
-	case "7d_api_cost", "all_time_api_cost", "5h_block_cost":
-		return false
-	}
-	// Prefixed tokens/cost metrics from providers.
-	if strings.HasPrefix(key, "tokens_today_") ||
-		strings.HasPrefix(key, "input_tokens_") ||
-		strings.HasPrefix(key, "output_tokens_") ||
-		strings.HasPrefix(key, "today_") ||
-		strings.HasPrefix(key, "7d_") ||
-		strings.HasPrefix(key, "all_time_") ||
-		strings.HasPrefix(key, "5h_block_") ||
-		strings.HasPrefix(key, "project_") ||
-		strings.HasPrefix(key, "agent_") {
-		return true
-	}
-	return false
-}
-
-// isCurrentStateMetric returns true for metrics that represent the current state
-// of a plan, billing cycle, or team — values that are always "latest" regardless
-// of time window. These are preserved even when stripping all-time cumulative
-// metrics for windowed views.
-func isCurrentStateMetric(key string) bool {
-	if strings.HasPrefix(key, "plan_") ||
-		strings.HasPrefix(key, "billing_") ||
-		strings.HasPrefix(key, "team_") ||
-		strings.HasPrefix(key, "spend_") ||
-		strings.HasPrefix(key, "individual_") {
-		return true
-	}
-	switch key {
-	case "today_cost",
-		"7d_api_cost", "all_time_api_cost", "5h_block_cost",
-		"usage_daily", "usage_weekly", "usage_five_hour":
-		return true
-	}
-	return false
-}
-
-func usageAuthoritativeCost(snap core.UsageSnapshot) float64 {
-	if m, ok := snap.Metrics["credit_balance"]; ok && m.Used != nil && *m.Used > 0 {
-		return *m.Used
-	}
-	if m, ok := snap.Metrics["spend_limit"]; ok && m.Used != nil && *m.Used > 0 {
-		return *m.Used
-	}
-	if m, ok := snap.Metrics["plan_total_spend_usd"]; ok && m.Used != nil && *m.Used > 0 {
-		return *m.Used
-	}
-	if m, ok := snap.Metrics["credits"]; ok && m.Used != nil && *m.Used > 0 {
-		return *m.Used
-	}
-	return 0
-}
-
-func sortedSeriesFromByDay(byDay map[string]float64) []core.TimePoint {
-	days := lo.Keys(byDay)
-	sort.Strings(days)
-
-	out := make([]core.TimePoint, 0, len(days))
-	for _, day := range days {
-		out = append(out, core.TimePoint{
-			Date:  day,
-			Value: byDay[day],
-		})
-	}
-	return out
 }
 
 // parseMCPToolName extracts server and function from an MCP tool name.
